requirements.txt
google-genai>=1.0.0
google-cloud-firestore>=2.16.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
python-dotenv>=1.0.0
